feat: Moved scene redraw responsibility to the renderer

+++ src/Renderer/Proxies/EllipseProxy.cpp
@@ -140,7 +140,7 @@ EllipseProxy::~EllipseProxy()
 
 }
 
-long long int & EllipseProxy::GetAssociatedModelId()
+size_t & EllipseProxy::GetAssociatedModelId()
 {
     return renderingModel->GetModelId();
 }
